rag-deepseek/pyproject.toml
[tool.poetry]
name = "rag-deepseek"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
faiss-cpu = "^1.10.0"
langchain-huggingface = "^0.1.2"
ollama = "^0.4.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
